fix: Stupid bug with long/int things that doesn't throw exceptions. :D

+++ Journal-Limpet.Shared/Models/UnsentJournalInfo.cs
@@ -6,12 +6,12 @@ namespace Journal_Limpet.Shared.Models
     public class UnsentJournalInfo
     {
         public Guid UserIdentifier { get; set; }
-        public int JournalCount { get; set; }
+        public long JournalCount { get; set; }
 
         public UnsentJournalInfo(DataRow row)
         {
             UserIdentifier = row.Field<Guid>("user_identifier");
-            JournalCount = row.Field<int>("journal_count");
+            JournalCount = row.Field<long>("journal_count");
         }
     }
 }
